requirements.txt
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dateparser==1.2.0
db-dtypes==1.2.0
fasteners==0.19
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.2.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
questionary @ git+https://github.com/FantasqueX/questionary@670653df2db3b71b2345db556bfb8040e0f20728
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rich-click==1.7.3
rsa==4.9
rtoml==0.10.0
six==1.16.0
sniffio==1.3.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
